fixed a few bugs and updated counter vec to double just in case that was causing problems with the math

+++ Analysis/include/utility.h
@@ -74,10 +74,10 @@ public:
     bool GetHist(TFile* f, TH1D* &h, TString string);
     // -------------------------------------------------------------------------
     // Function to tabulate all the nuetrino types and flavours
-    void Tabulate(std::string interaction, std::string classification, int type, std::vector<int> &counter_v);
+    void Tabulate(std::string interaction, std::string classification, int type, std::vector<double> &counter_v);
     // -------------------------------------------------------------------------
     // Function to print the tabulated events
-    void PrintInfo(std::vector<int> counter_v, double intime_scale_factor, double data_scale_factor, double dirt_scale_factor, std::string cut_name, int tot_true_infv_nues);
+    void PrintInfo(std::vector<double> counter_v, double intime_scale_factor, double data_scale_factor, double dirt_scale_factor, std::string cut_name, int tot_true_infv_nues);
     // -------------------------------------------------------------------------
 
     // Other definitions for code
